[ refactor ] 將 redux 重構為 slice 寫法 5
